Update FinalMarketplace.sol
Contract name changed to `FinalMarketplace`.
The array of collections was not being populated, because the only way to populate it is by calling the `addCollection` function and it's not being called in my tests. Due to the time constraints we have, I commented the lines related to that array inside the constructor.